+++ ExperimentalEnemyInteractions/EnemyPatches/SandSpiderAIPatch.cs
@@ -118,7 +118,7 @@ namespace NaturalSelection.EnemyPatches
                         }
                         else
                         {
-                            spiderData.closestEnemy = LibraryCalls.FindClosestEnemy(ref tempList, spiderData.closestEnemy, __instance, usePathLenghtAsDistance: true);
+                            spiderData.closestEnemy = LibraryCalls.FindClosestEnemy(ref tempList, spiderData.closestEnemy, __instance, usePathLenghtAsDistance: Script.usePathToFindClosestEnemy);
                         }
 
                         if (spiderData.closestEnemy != null && __instance.CheckLineOfSightForPosition(spiderData.closestEnemy.transform.position, 80f, 15, 2f, __instance.eye) != false && !spiderData.closestEnemy.isEnemyDead)