+++ ExperimentalEnemyInteractions/Generics/MyModConfig.cs
@@ -6,6 +6,10 @@ using System.Reflection;
 namespace NaturalSelection.Generics;
     class MyModConfig
     {
+    //experimental
+    public readonly ConfigEntry<bool> useExperimentalCoroutines;
+    public readonly ConfigEntry<bool> usePathfindinglibCoroutines;
+    public readonly ConfigEntry<bool> usePathToFindClosestEnemy;
     //experimental fixes
     public readonly ConfigEntry<bool> sandwormCollisionOverride;
     public readonly ConfigEntry<float> blobAIOpeningDoorsMultiplier;
@@ -15,7 +19,6 @@ namespace NaturalSelection.Generics;
     public readonly ConfigEntry<float> agentRadiusModifier;
     public readonly ConfigEntry<float> globalListsUpdateInterval;
     public readonly ConfigEntry<string> customSizeOverrideList;
-    public readonly ConfigEntry<bool> useExperimentalCoroutines;
     //enemy bools
     public readonly ConfigEntry<bool> enableSpider;
     public readonly ConfigEntry<bool> enableSlime;
@@ -83,8 +86,9 @@ namespace NaturalSelection.Generics;
         {
             //experimental fixes
             sandwormCollisionOverride = cfg.Bind("Experimental Fixes", "Sandworm collision override", false, "Override vanilla sandworm collisions. May fix lag when sandworm collides with multiple enemies at once. \n \n May be removed in the future.");
-            useExperimentalCoroutines = cfg.Bind("Experimental", "Enable experimental library coroutines", false, "Use experimental coroutines for finding closest enemy. May improve performance. Limited testing.");
-
+            useExperimentalCoroutines = cfg.Bind("Experimental", "Enable experimental library coroutines", false, "Use experimental coroutines for finding closest enemy. May improve performance.");
+            usePathfindinglibCoroutines = cfg.Bind("Experimental", "Enable pathfindinglib coroutines", false, "Use experimental pathfindinglib coroutines.");
+            usePathToFindClosestEnemy = cfg.Bind("Experimental", "Use paths to find closest enemy", false, "Enemies will use pathfinding to find closest enemy. May be performance heavy.");
             //general settings
             stableMode = cfg.Bind("General Settings", "Toggle stable mode", true, "When true, the mod will exlude patches that are WIP or are experimental from loading. Requires restart.");
             IgnoreImmortalEnemies = cfg.Bind("General Settings", "Ignore Immortal Enemies", false, "All immortal enemies will be ignored by majority of entities.");